refactoring of files (conversion_handler + solving_state)
