+++ src/smt/CMakeLists.txt
@@ -75,6 +75,7 @@ z3_add_component(smt
     theory_str_noodler/theory_str_noodler_model.cpp
     theory_str_noodler/formula_preprocess.cpp
     theory_str_noodler/inclusion_graph.cc
+    theory_str_noodler/solving_state.cpp
     theory_str_noodler/decision_procedure.cpp
     theory_str_noodler/nielsen_decision_procedure.cpp
     theory_str_noodler/memb_heuristics_procedures.cpp
@@ -89,6 +90,7 @@ z3_add_component(smt
     theory_str_noodler/counter_automaton.cpp
     theory_str_noodler/ca_str_constr.cpp
     theory_str_noodler/unary_decision_procedure.cpp
+    theory_str_noodler/conversion_handler.cpp
     theory_str_mc.cpp
     theory_str_regex.cpp
     theory_user_propagator.cpp
