+++ src/smt/theory_str_noodler/formula_preprocess.h
@@ -72,6 +72,22 @@ namespace smt::noodler {
         return true;
     }
 
+    template<typename T>
+    bool set_disjoint(const std::set<T>& t1, const std::set<T>& t2) {
+        if (t1.size() < t2.size()) {
+            for(const auto& t : t1) {
+                if(t2.contains(t))
+                    return false;
+            }
+        } else {
+            for(const auto& t : t2) {
+                if(t1.contains(t))
+                    return false;
+            }
+        }
+        return true;
+    }
+
     template<typename T, typename S, typename P>
     std::set<S> set_map(const std::set<T>& st, P pred) {
         std::set<S> ret;
@@ -335,7 +351,7 @@ namespace smt::noodler {
 
         LenNode len_formula;
         std::unordered_set<BasicTerm> len_variables;
-        std::unordered_set<BasicTerm> conversion_vars; // all conversion vars should always be also in len vars, also it should not contain literals
+        std::set<BasicTerm> conversion_vars; // all conversion vars should always be also in len vars, also it should not contain literals
 
         const theory_str_noodler_params& m_params;
 
@@ -365,7 +381,7 @@ namespace smt::noodler {
 
 
     public:
-        FormulaPreprocessor(Formula conj, AutAssignment ass, std::unordered_set<BasicTerm> lv, const theory_str_noodler_params &par, std::unordered_set<BasicTerm> conversion_vars) :
+        FormulaPreprocessor(Formula conj, AutAssignment ass, std::unordered_set<BasicTerm> lv, const theory_str_noodler_params &par, std::set<BasicTerm> conversion_vars) :
             formula(conj),
             fresh_var_cnt(0),
             aut_ass(ass),
@@ -404,7 +420,7 @@ namespace smt::noodler {
         void infer_alignment();
         void common_prefix_propagation();
         void common_suffix_propagation();
-        void conversions_validity(std::vector<TermConversion>& conversions);
+        void conversions_validity(const std::vector<TermConversion>& conversions);
         bool has_unsat_transducers();
 
         void underapprox_var_language(const BasicTerm& var);
